PAYARA-3864 reorganised packages to be consistent between client and endpoint, added protocol to client
