+++ appserver/ejb/ejb-http-remoting/client/pom.xml
@@ -163,5 +163,12 @@
             <artifactId>yasson</artifactId>
             <scope>runtime</scope>
         </dependency>
+
+        <dependency>
+            <groupId>fish.payara.ejb.http</groupId>
+            <artifactId>ejb-http-endpoint</artifactId>
+            <version>${project.version}</version>            
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
