+++ appserver/ejb/ejb-http-remoting/client/src/main/java/fish/payara/ejb/http/client/MediaTypes.java
@@ -38,7 +38,7 @@
  *    holder.
  */
 
-package fish.payara.ejb;
+package fish.payara.ejb.http.client;
 
 import javax.ws.rs.core.MediaType;
