+++ appserver/ejb/ejb-http-remoting/client/src/main/java/fish/payara/ejb/http/client/protocol/InvokeMethodResponse.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.ejb.http.protocol;
+package fish.payara.ejb.http.client.protocol;
 
 import javax.json.bind.annotation.JsonbPropertyOrder;
 import javax.json.bind.annotation.JsonbTypeDeserializer;
