+++ appserver/ejb/ejb-http-remoting/endpoint/src/main/java/fish/payara/ejb/http/endpoint/protocol/rs/JsonbLookupMessageBodyReader.java
@@ -37,12 +37,48 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.ejb.endpoint;
+package fish.payara.ejb.http.endpoint.protocol.rs;
 
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
+import java.io.IOException;
+import java.io.InputStream;
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Type;
 
-@ApplicationPath("/")
-public class EjbOverHttpApplication extends Application {
-    // required to trigger JAX-RS
+import javax.json.Json;
+import javax.json.JsonObject;
+import javax.json.JsonReader;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.MultivaluedMap;
+import javax.ws.rs.ext.MessageBodyReader;
+import javax.ws.rs.ext.Provider;
+
+import fish.payara.ejb.http.endpoint.MediaTypes;
+import fish.payara.ejb.http.endpoint.protocol.LookupRequest;
+
+/**
+ * Reads the {@link LookupRequest} in case of JSONB serialisation.
+ *
+ * @author Jan Bernitt
+ */
+@Provider
+@Consumes(MediaTypes.JSON)
+public class JsonbLookupMessageBodyReader implements MessageBodyReader<LookupRequest> {
+
+    @Override
+    public boolean isReadable(Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType) {
+        return LookupRequest.class.isAssignableFrom(type);
+    }
+
+    @Override
+    public LookupRequest readFrom(Class<LookupRequest> type, Type genericType, Annotation[] annotations,
+            MediaType mediaType, MultivaluedMap<String, String> httpHeaders, InputStream entityStream)
+            throws IOException, WebApplicationException {
+        try (JsonReader jsonReader = Json.createReader(entityStream)) {
+            JsonObject request = jsonReader.readObject();
+            return new LookupRequest(request.getString("lookup"));
+        }
+    }
+
 }
